requirements.txt
flask
genai
MimeTypes
Optional
google-generativeai
google-cloud-vision
google-genai
pymongo